# openspec/changes/sdlc-visualizer/interface-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Arsitect SDLC Visualizer API
  version: 1.0.0
  description: Arsitect 可视化驾驶舱 REST API 契约（基于详细设计 20/21 模块导出）
servers:
- url: /api/v1
  description: API v1 基线
paths:
...
  /monitoring/projects/{project_id}/stats:
    get:
      tags:
      - monitoring
      summary: Get Project Stats
      description: Get per-project monitoring stats.
      operationId: get_project_stats_api_v1_monitoring_projects__project_id__stats_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStatsDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /monitoring/projects/{project_id}/operation-logs:
    get:
      tags:
      - monitoring
      summary: List Operation Logs
      description: List operation logs for a project.
      operationId: list_operation_logs_api_v1_monitoring_projects__project_id__operation_logs_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationLogListDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /projects/{project_id}:
    get:
      tags:
      - projects
      summary: Get Project
      description: Get project details.
      operationId: get_project_api_v1_projects__project_id__get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - projects
      summary: Update Project
      description: Update project info.
      operationId: update_project_api_v1_projects__project_id__patch
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateDTO'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{project_id}/archive:
    post:
      tags:
      - projects
      summary: Archive Project
      description: Archive a project.
      operationId: archive_project_api_v1_projects__project_id__archive_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Archive Project Api V1 Projects  Project Id  Archive Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{project_id}/activate:
    post:
      tags:
      - projects
      summary: Activate Project
      description: Activate (confirm) a project.
      operationId: activate_project_api_v1_projects__project_id__activate_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{project_id}/cancel:
    post:
      tags:
      - projects
      summary: Cancel Project
      description: Cancel a project.
      operationId: cancel_project_api_v1_projects__project_id__cancel_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OperationLogDTO:
      properties:
        log_id:
          type: string
          title: Log Id
        project_id:
          type: string
          title: Project Id
        operator_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Operator Id
        action:
          type: string
          title: Action
        target_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Type
        target_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Id
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - log_id
      - project_id
      - operator_id
      - action
      - target_type
      - target_id
      - detail
      - created_at
      title: OperationLogDTO
      description: Operation log entry response.
    OperationLogListDTO:
      properties:
        logs:
          items:
            $ref: '#/components/schemas/OperationLogDTO'
          type: array
          title: Logs
        total:
          type: integer
          title: Total
      type: object
      required:
      - logs
      - total
      title: OperationLogListDTO
      description: Paginated operation log list response.
...
    ProjectResponseDTO:
      properties:
        project_id:
          type: string
          title: Project Id
        project_name:
          type: string
          title: Project Name
        project_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Description
        project_status:
          type: string
          title: Project Status
        application_id:
          type: string
          title: Application Id
        template_level:
          type: string
          title: Template Level
        progress_percent:
          type: integer
          title: Progress Percent
        current_stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Current Stage
        risk_level:
          type: string
          title: Risk Level
        last_activity_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Activity At
        last_activity_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Activity Type
        size_estimate_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Size Estimate Id
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - project_id
      - project_name
      - project_description
      - project_status
      - application_id
      - template_level
      - progress_percent
      - current_stage
      - risk_level
      - last_activity_type
      - size_estimate_id
      title: ProjectResponseDTO
      description: DTO for project response.
...
    ProjectStatsDTO:
      properties:
        stage_count:
          type: integer
          title: Stage Count
        execution_count:
          type: integer
          title: Execution Count
        gate_count:
          type: integer
          title: Gate Count
        log_count:
          type: integer
          title: Log Count
      type: object
      required:
      - stage_count
      - execution_count
      - gate_count
      - log_count
      title: ProjectStatsDTO
      description: Per-project monitoring stats response.
    ProjectUpdateDTO:
      properties:
        project_name:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Project Name
        project_description:
          anyOf:
          - type: string
            maxLength: 256
          - type: 'null'
          title: Project Description
      type: object
      title: ProjectUpdateDTO
      description: DTO for updating a project.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: MVP 阶段为本地会话占位，P1 升级为 JWT
...